requirements.txt
PyYAML
beautifulsoup4==4.7.1
beautifultable==0.7.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
pymongo==3.8.0
requests==2.22.0
soupsieve==1.9.1
urllib3==1.25.3
